PCI: fix Sapphire PCI rebar quirk
There was a quirk added to add a workaround for a Sapphire
RX 5600 XT Pulse.  However, the quirk only checks the vendor
ids and not the subsystem ids.  The quirk really should
have checked the subsystem vendor and device ids as now
this quirk gets applied to all RX 5600 and it seems to
cause problems on some Dell laptops.  Add a subsystem vendor
id check to limit the quirk to Sapphire boards.

Closes: https://gitlab.freedesktop.org/drm/amd/-/issues/1707
Fixes: 907830b0fc9e374 ("PCI: Add a REBAR size quirk for Sapphire RX 5600 XT Pulse")
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Cc: Christian König <christian.koenig@amd.com>
Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: Nirmoy Das <nirmoy.aiemd@gmail.com>